Added resource saving options
This commit adds 2 new options "--auto-pause/-p" and "--auto-stop/-s"
- By using a Wayland feature called "surface frame callback"
mpvpaper can detect when the wallpaper is not visible.
- With this feature mpvpaper can pause/stop automagically
while the wallpaper can't be seen.

The pause option will literally just pause mpv.
- In order to save processing power more seamlessly.

While the stop option will sort of "suspend" mpvpaper
to a separate holder program call "mpvpaper-holder"
- This reduces not just processing power, but also RAM usage
at the cost of being at bit more obvious when it does stop.

Along with the automagical ways to pause/stop.
There also now a "pauselist" and "stoplist".
Which will pause/stop if a program is found running in it's list.
- To use, simply create these text based lists in "~/.config/mpvpaper/"
with the file name "pauselist" and/or "stoplist" respectively.
- Then add programs that can be found with the "pidof" command.
- Make sure the programs are separated by either spaces or newlines.
Like so: "wine firefox steam"

This commit also includes some minor fixes to many too name.
Such as converting some ints to bools
and changing the order of loading configs.