Add ALL as a new way to select all outputs
This will replace the old "*" or '*' way of selecting all outputs.
Of course in order to not needlessly break old configurations
"*" will remain in the code, but not in the documentation.

The reason for adding ALL is to avoid further confusion in the future.
As * in most shells is a wildcard and might not be obvious that
* needs to be used in quotes to avoid calling the wildcard behavior.